Move all image processing logic into utils, and expose publically
